Add Buildage submodule
This involves making copies of the Buildage sample input files and customizing them per the Getting Started instructions as well as modifying the solution files to use the new generated GlobalAssemblyInfo.cs.
Since the solution already had a GlobalAssemblyInfo.cs, it was removed since it no longer needs to be kept under source control.